chore: clippy 清理(48→0 项目 lint)
- 自动修复 36 项(unnecessary min/max、redundant closure、io::Error::other、
  needless_borrow、is_multiple_of、manual ok 等)
- 手动修复 10 项:
  - skill_updates redundant guard → or-pattern Err(Http(401|403|429))
  - streaming_message let-else → ? operator
  - skill_lock annotate &mut Vec → &mut [SkillInfo]
  - scope unwrap after is_some → if let pattern
  - title identical if/else → 合并 json/text 分支
  - changes identical if/else → 折叠
  - GIT_TIMEOUT_SECS 标 #[allow(dead_code)](预留宿主接线)
  - request_security is_api_request_allowed #[allow(too_many_arguments)](8 参数对齐 TS Request 字段)
  - catalog sort_by 三处 #[allow(unnecessary_sort_by)](tuple/Vec 长度比较)
  - fs/mod.rs 中文段误判为 doc list → suppress

254 tests passed, 0 clippy warnings.

+++ src/session/entries.rs
@@ -109,7 +109,7 @@ fn parse_iso8601_millis(value: &str) -> Option<i64> {
     let utc_hour = hour as i32 - utc_hour;
     let utc_minute = minute as i32 - utc_minute;
 
-    let days = days_from_civil(year as i64, month as u32, day as u32)?;
+    let days = days_from_civil(year as i64, month, day)?;
     days.checked_mul(86_400_000)?
         .checked_add((utc_hour as i64 * 3600 + utc_minute as i64 * 60 + second as i64) * 1000)?
         .checked_add(millis as i64)
